move app out of package; rename app
